moved create_endpoint to endpoints.py
