+++ calm/dsl/cli/endpoint_commands.py
@@ -1,15 +1,11 @@
-import sys
 import click
-import json
 
-from calm.dsl.api import get_api_client
-from calm.dsl.config import get_config
 from calm.dsl.tools import get_logging_handle
 
 from .main import compile, get, describe, delete, create, format
 from .endpoints import (
     get_endpoint_list,
-    compile_endpoint,
+    create_endpoint_command,
     delete_endpoint,
     describe_endpoint,
     format_endpoint_command,
@@ -40,61 +36,6 @@ def _get_endpoint_list(name, filter_by, limit, offset, quiet, all_items):
     get_endpoint_list(name, filter_by, limit, offset, quiet, all_items)
 
 
-def create_endpoint(
-    client, endpoint_payload, name=None, description=None, force_create=False
-):
-
-    endpoint_payload.pop("status", None)
-
-    if name:
-        endpoint_payload["spec"]["name"] = name
-        endpoint_payload["metadata"]["name"] = name
-
-    if description:
-        endpoint_payload["spec"]["description"] = description
-
-    endpoint_resources = endpoint_payload["spec"]["resources"]
-    endpoint_name = endpoint_payload["spec"]["name"]
-    endpoint_desc = endpoint_payload["spec"]["description"]
-
-    return client.endpoint.upload_with_secrets(
-        endpoint_name, endpoint_desc, endpoint_resources, force_create=force_create,
-    )
-
-
-def create_endpoint_from_json(
-    client, path_to_json, name=None, description=None, force_create=False
-):
-
-    endpoint_payload = json.loads(open(path_to_json, "r").read())
-    return create_endpoint(
-        client,
-        endpoint_payload,
-        name=name,
-        description=description,
-        force_create=force_create,
-    )
-
-
-def create_endpoint_from_dsl(
-    client, endpoint_file, name=None, description=None, force_create=False
-):
-
-    endpoint_payload = compile_endpoint(endpoint_file)
-    if endpoint_payload is None:
-        err_msg = "User endpoint not found in {}".format(endpoint_file)
-        err = {"error": err_msg, "code": -1}
-        return None, err
-
-    return create_endpoint(
-        client,
-        endpoint_payload,
-        name=name,
-        description=description,
-        force_create=force_create,
-    )
-
-
 @create.command("endpoint", feature_min_version="3.0.0")
 @click.option(
     "--file",
@@ -113,74 +54,10 @@ def create_endpoint_from_dsl(
     default=False,
     help="Deletes existing blueprint with the same name before create.",
 )
-def create_endpoint_command(endpoint_file, name, description, force):
+def _create_endpoint_command(endpoint_file, name, description, force):
     """Creates a endpoint"""
 
-    client = get_api_client()
-
-    if endpoint_file.endswith(".json"):
-        res, err = create_endpoint_from_json(
-            client,
-            endpoint_file,
-            name=name,
-            description=description,
-            force_create=force,
-        )
-    elif endpoint_file.endswith(".py"):
-        res, err = create_endpoint_from_dsl(
-            client,
-            endpoint_file,
-            name=name,
-            description=description,
-            force_create=force,
-        )
-    else:
-        LOG.error("Unknown file format {}".format(endpoint_file))
-        return
-
-    if err:
-        LOG.error(err["error"])
-        return
-
-    endpoint = res.json()
-    endpoint_uuid = endpoint["metadata"]["uuid"]
-    endpoint_name = endpoint["metadata"]["name"]
-    endpoint_status = endpoint.get("status", {})
-    endpoint_state = endpoint_status.get("state", "DRAFT")
-    LOG.debug("Endpoint {} has state: {}".format(endpoint_name, endpoint_state))
-
-    if endpoint_state != "ACTIVE":
-        msg_list = endpoint_status.get("message_list", [])
-        if not msg_list:
-            LOG.error("Endpoint {} created with errors.".format(endpoint_name))
-            LOG.debug(json.dumps(endpoint_status))
-            sys.exit(-1)
-
-        msgs = []
-        for msg_dict in msg_list:
-            msgs.append(msg_dict.get("message", ""))
-
-        LOG.error(
-            "Endpoint {} created with {} error(s): {}.".format(
-                endpoint_name, len(msg_list), msgs
-            )
-        )
-        sys.exit(-1)
-
-    LOG.info("Endpoint {} created successfully.".format(endpoint_name))
-    config = get_config()
-    pc_ip = config["SERVER"]["pc_ip"]
-    pc_port = config["SERVER"]["pc_port"]
-    link = "https://{}:{}/console/#page/explore/calm/endpoints/{}".format(
-        pc_ip, pc_port, endpoint_uuid
-    )
-
-    stdout_dict = {
-        "name": endpoint_name,
-        "link": link,
-        "state": endpoint_state,
-    }
-    click.echo(json.dumps(stdout_dict, indent=4, separators=(",", ": ")))
+    create_endpoint_command(endpoint_file, name, description, force)
 
 
 @delete.command("endpoint", feature_min_version="3.0.0")
